Guts the Logger
The Logger idea didn't make sense because each time this routing
application runs, it runs in a unique process. The logs from previous
requests will not persist because from "this instance" of the appliation
running, those requests never occurred.

So, this commit reverts back to the ProtectedResource builder just
returning the body of the resource (which, hey! makes sense). Move the
logging details into the server's logging functionality.